Squash data contract versions
While developing data contracts we've mostly been making in-place breaking changes on v1
contracts but for a few we added multiple versions before our development time versioning
policy was decided on. As preparation to lock down contracts we are simplifying by getting
rid of contract versions that only existed during unsupported development builds. Going
forward we don't expect a numbering reset like this to happen again.
