+++ docs/design/datacontracts/data-descriptor-meanings.json
@@ -310,7 +310,7 @@
     "InterpMethodContextFrame.Stack": "Pointer to the stack base for this interpreted method, used as the frame pointer when interpreter GC info uses a stack-base register",
     "InterpreterFrame.IsFaulting": "Boolean indicating whether the topmost interpreted frame has thrown an exception. When set, the context for the top interpreted frame must include CONTEXT_EXCEPTION_ACTIVE so exception unwinders treat the IP as a faulting instruction rather than a return-from-call",
     "InterpreterFrame.TopInterpMethodContextFrame": "Pointer to the InterpreterFrame's top InterpMethodContextFrame",
-    "InterpreterPrecodeData.ByteCodeAddr": "pointer to the InterpByteCodeStart for the interpreter bytecode (Version 3 only)",
+    "InterpreterPrecodeData.ByteCodeAddr": "pointer to the InterpByteCodeStart for the interpreter bytecode",
     "InterpreterRealCodeHeader.DebugInfo": "Pointer to the DebugInfo for interpreter code",
     "InterpreterRealCodeHeader.GCInfo": "Pointer to the GCInfo encoding for interpreter code",
     "InterpreterRealCodeHeader.JitEHInfo": "Pointer to the `EE_ILEXCEPTION` containing exception clauses for interpreter code",
@@ -449,20 +449,17 @@
     "PlatformMetadata.CodePointerFlags": "fields describing the behavior of target code pointers",
     "PlatformMetadata.PrecodeMachineDescriptor": "precode stub-related platform specific properties",
     "PortableEntryPoint.MethodDesc": "Method desc of portable entrypoint (only defined if `FeaturePortableEntrypoints` is enabled)",
-    "PrecodeMachineDescriptor.FixupBytes": "Assembly code of a FixupStub (Version 3 only)",
-    "PrecodeMachineDescriptor.FixupIgnoredBytes": "Bytes to ignore of when comparing FixupBytes to an actual block of memory in the target process. (Version 3 only)",
+    "PrecodeMachineDescriptor.FixupBytes": "Assembly code of a FixupStub",
+    "PrecodeMachineDescriptor.FixupIgnoredBytes": "Bytes to ignore when comparing FixupBytes to an actual block of memory in the target process.",
     "PrecodeMachineDescriptor.FixupPrecodeType": "Precode type byte for a fixup precode",
-    "PrecodeMachineDescriptor.FixupStubPrecodeSize": "Byte size of FixupBytes and FixupIgnoredBytes (Version 3 only)",
+    "PrecodeMachineDescriptor.FixupStubPrecodeSize": "Byte size of FixupBytes and FixupIgnoredBytes",
     "PrecodeMachineDescriptor.DynamicHelperPrecodeType": "Precode type byte for a dynamic helper precode",
     "PrecodeMachineDescriptor.InterpreterPrecodeType": "Precode type byte for an interpreter precode",
-    "PrecodeMachineDescriptor.OffsetOfPrecodeType": "See ReadPrecodeType (Version 1 and 2 only)",
     "PrecodeMachineDescriptor.PInvokeImportPrecodeType": "Precode type byte for a P/Invoke import precode",
-    "PrecodeMachineDescriptor.ReadWidthOfPrecodeType": "See ReadPrecodeType  (Version 1 and 2 only)",
-    "PrecodeMachineDescriptor.ShiftOfPrecodeType": "See ReadPrecodeType  (Version 1 and 2 only)",
-    "PrecodeMachineDescriptor.StubBytes": "Assembly code of a StubPrecode (Version 3 only)",
+    "PrecodeMachineDescriptor.StubBytes": "Assembly code of a StubPrecode",
     "PrecodeMachineDescriptor.StubCodePageSize": "Size of a precode code page (in bytes)",
-    "PrecodeMachineDescriptor.StubIgnoredBytes": "Bytes to ignore of when comparing StubBytes to an actual block of memory in the target process. (Version 3 only)",
-    "PrecodeMachineDescriptor.StubPrecodeSize": "Byte size of StubBytes and StubIgnoredBytes (Version 3 only)",
+    "PrecodeMachineDescriptor.StubIgnoredBytes": "Bytes to ignore when comparing StubBytes to an actual block of memory in the target process.",
+    "PrecodeMachineDescriptor.StubPrecodeSize": "Byte size of StubBytes and StubIgnoredBytes",
     "PrecodeMachineDescriptor.StubPrecodeType": "precode sort byte for stub precodes",
     "PrecodeMachineDescriptor.ThisPointerRetBufPrecodeType": "Precode type byte for a this-pointer return-buffer precode",
     "PrecodeMachineDescriptor.UMEntryPrecodeType": "Precode type byte for a UMEntry precode",
@@ -586,8 +583,7 @@
     "StubDispatchFrame.MethodDescPtr": "Pointer to Frame's method desc",
     "StubDispatchFrame.RepresentativeMTPtr": "Pointer to Frame's method table pointer",
     "StubDispatchFrame.RepresentativeSlot": "Frame's method table slot",
-    "StubPrecodeData.MethodDesc": "pointer to the MethodDesc associated with this stub precode (Version 1 only)",
-    "StubPrecodeData.SecretParam": "pointer to the MethodDesc associated with this stub precode or a second stub data pointer for other types (Version 2 only)",
+    "StubPrecodeData.SecretParam": "pointer to the MethodDesc associated with this stub precode or a second stub data pointer for other types",
     "StubPrecodeData.Type": "precise sort of stub precode",
     "SyncBlock.EnCInfo": "Pointer to Edit-and-Continue added-field information for the object; optional when Edit and Continue is not configured",
     "SyncBlock.HashCode": "Hash code stored in the sync block",
@@ -629,7 +625,7 @@
     "TailCallFrame.CalleeSavedRegisters": "Address of the embedded nonvolatile-register values saved in the tailcall frame",
     "TailCallFrame.ReturnAddress": "Return address saved in the tailcall frame",
     "TailCallFrame.Size": "Size in bytes of the tailcall frame, used to restore the caller's stack pointer",
-    "ThisPtrRetBufPrecodeData.MethodDesc": "pointer to the MethodDesc associated with the ThisPtrRetBufPrecode (Version 2 only)",
+    "ThisPtrRetBufPrecodeData.MethodDesc": "pointer to the MethodDesc associated with the ThisPtrRetBufPrecode",
     "Thread.CachedStackBase": "Pointer to the base of the stack",
     "Thread.CachedStackLimit": "Pointer to the limit of the stack",
     "Thread.CurrentCustomDebuggerNotification": "Handle to the current custom debugger notification object",
