+++ src/coreclr/nativeaot/Runtime/datadescriptor/datadescriptor.inc
@@ -252,7 +252,7 @@ CDAC_GLOBAL_CONTRACT(Exception, n1)
 CDAC_GLOBAL_CONTRACT(Object, n1)
 CDAC_GLOBAL_CONTRACT(RuntimeTypeSystem, n1)
 CDAC_GLOBAL_CONTRACT(RuntimeInfo, c1)
-CDAC_GLOBAL_CONTRACT(StressLog, c2)
+CDAC_GLOBAL_CONTRACT(StressLog, c1)
 CDAC_GLOBAL_CONTRACT(Loader, n1)
 CDAC_GLOBAL_CONTRACT(ExecutionManager, n1)
 CDAC_GLOBAL_CONTRACT(SyncBlock, n1)
