+++ src/native/managed/cdac/Microsoft.Diagnostics.DataContractReader.Contracts/Data/StubPrecodeData.cs
@@ -6,13 +6,6 @@ namespace Microsoft.Diagnostics.DataContractReader.Data;
 [CdacType(nameof(DataType.StubPrecodeData))]
 internal sealed partial class StubPrecodeData_1 : IData<StubPrecodeData_1>
 {
-    [Field] public partial TargetPointer MethodDesc { get; }
-    [Field] public partial byte Type { get; }
-}
-
-[CdacType(nameof(DataType.StubPrecodeData))]
-internal sealed partial class StubPrecodeData_2 : IData<StubPrecodeData_2>
-{
     [Field] public partial TargetPointer SecretParam { get; }
     [Field] public partial byte Type { get; }
 }
