+++ src/native/managed/cdac/tests/UnitTests/MockDescriptors/MockDescriptors.ExecutionManager.cs
@@ -748,7 +748,6 @@ internal sealed class MockExecutionManagerBuilder
         NibbleMapTestBuilderBase nibBuilder = Version switch
         {
             "c1" => new NibbleMapTestBuilder_1(codeRangeStart, codeRangeSize, _nibbleMapAllocator, Builder.TargetTestHelpers.Arch),
-            "c2" => new NibbleMapTestBuilder_2(codeRangeStart, codeRangeSize, _nibbleMapAllocator, Builder.TargetTestHelpers.Arch),
             _ => throw new InvalidOperationException($"Unknown version '{Version}'"),
         };
